+++ changelog.md
@@ -17,6 +17,8 @@ Given a version number MAJOR.MINOR.PATCH:
  - Load google fonts with https
  - Only accessible reports (that match the organization name) are displayed in the menu
  - Issue 252: added picture notes
+ - Issues 119 and 323: 'Measured applied" added everywhere where a measure has been applied
+ - Fixed bug: when adding a new heating system as a measure the id of the heating system was wrong
 
 
 ## Version 1.1.0